requirements.txt
asgiref==3.2.3
bson==0.5.8
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dataclasses==0.6
Django==2.2.10
django-cors-headers==3.2.1
django-logpipe==1.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djongo==1.3.1
httpie==2.0.0
idna==2.9
kafka-python==2.0.1
lru-dict==1.1.6
Pillow==2.6.1
Pygments==2.5.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
